+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Edit internet search instructions in settings.
 - File size displayed
 - LICENSE: AGPL v3
+- Option (on by default) to use the category in tags for file uploads.
 
 ### Changed